# pipe: enable writev for pipe handles on Unix