@aqrln 2017-01-08
I ran node_redis benchmark and this patch indeed seems to fix #5095.

Full results: https://gist.github.com/aqrln/6fb1af97459e207ec654e75a6fc4cd00

Comparison of benchmarks before and after the patch on Linux:

```
Comparing before, linux-before.txt ( 37 lines) to after, linux-after.txt ( 37 lines)
clients: 1:        0 ->      0 ops/sec (∆ -      0 -    NaN%)
PING:          27089 ->  30315 ops/sec (∆ +   3226 +  11.91%)
PING:         618513 -> 668113 ops/sec (∆ +  49600 +   8.02%)
SET 4B str:    25407 ->  26432 ops/sec (∆ +   1025 +   4.03%)
SET 4B str:   432767 -> 446921 ops/sec (∆ +  14154 +   3.27%)
SET 4B buf:    12026 ->  12261 ops/sec (∆ +    235 +   1.95%)
SET 4B buf:    17086 -> 171152 ops/sec (∆ + 154066 + 901.71%)
GET 4B str:    28040 ->  28977 ops/sec (∆ +    937 +   3.34%)
GET 4B str:   507737 -> 525030 ops/sec (∆ +  17293 +   3.41%)
GET 4B buf:    23984 ->  27532 ops/sec (∆ +   3548 +  14.79%)
GET 4B buf:   454798 -> 523850 ops/sec (∆ +  69052 +  15.18%)
SET 4KiB str:  20527 ->  23892 ops/sec (∆ +   3365 +  16.39%)
SET 4KiB str: 140304 -> 154618 ops/sec (∆ +  14314 +  10.20%)
SET 4KiB buf:  10802 ->  11967 ops/sec (∆ +   1165 +  10.79%)
SET 4KiB buf:  15961 -> 129128 ops/sec (∆ + 113167 + 709.02%)
GET 4KiB str:  23923 ->  25663 ops/sec (∆ +   1740 +   7.27%)
GET 4KiB str: 171351 -> 187105 ops/sec (∆ +  15754 +   9.19%)
GET 4KiB buf:  23174 ->  26569 ops/sec (∆ +   3395 +  14.65%)
GET 4KiB buf: 144582 -> 163794 ops/sec (∆ +  19212 +  13.29%)
INCR:          25148 ->  27120 ops/sec (∆ +   1972 +   7.84%)
INCR:         471591 -> 534386 ops/sec (∆ +  62795 +  13.32%)
LPUSH:         24301 ->  26746 ops/sec (∆ +   2445 +  10.06%)
LPUSH:        383287 -> 441743 ops/sec (∆ +  58456 +  15.25%)
LRANGE 10:     23138 ->  25018 ops/sec (∆ +   1880 +   8.13%)
LRANGE 10:    204398 -> 236106 ops/sec (∆ +  31708 +  15.51%)
LRANGE 100:    15573 ->  17351 ops/sec (∆ +   1778 +  11.42%)
LRANGE 100:    33627 ->  37870 ops/sec (∆ +   4243 +  12.62%)
SET 4MiB str:    196 ->    215 ops/sec (∆ +     19 +   9.69%)
SET 4MiB str:    193 ->    221 ops/sec (∆ +     28 +  14.51%)
SET 4MiB buf:    398 ->    461 ops/sec (∆ +     63 +  15.83%)
SET 4MiB buf:    415 ->    468 ops/sec (∆ +     53 +  12.77%)
GET 4MiB str:    415 ->    466 ops/sec (∆ +     51 +  12.29%)
GET 4MiB str:    211 ->    227 ops/sec (∆ +     16 +   7.58%)
GET 4MiB buf:    295 ->    321 ops/sec (∆ +     26 +   8.81%)
GET 4MiB buf:    296 ->    318 ops/sec (∆ +     22 +   7.43%)
:              85250 ->  85161 ops/sec (∆ -     89 -   0.10%)
Mean difference in ops/sec:  +18075.4
```