@trevnorris 2017-01-09
@aqrln you mentioned in IRC that this fixes an issue I self assigned. If that's the case, mind adding the `Fixes:` field in the commit message?