@jasnell 2017-01-10
CI: https://ci.nodejs.org/job/node-test-pull-request/5774/