@trevnorris 2017-01-11
@aqrln Sorry for being pedantic but the full URL should be used. not just the issue number. other than that, LGTM