@aqrln 2017-01-11
@trevnorris thanks for pointing that out, I've updated the commit message.