@jasnell 2017-01-12
Landed in d85d120.
Fixed commit log to use `src:` subsystem rather than `pipe:` (which is not an actual component