@Salakar 2017-01-16
@aqrln awesome, thanks for this!