@jasnell 2017-03-08
It should be fine. From what I understand it went out in 7.5 and there have been no regressions that I've seen.